+++ code/cell.modules/pkg.sys/net/src/common/libs.ts
@@ -37,6 +37,7 @@ export { PropList, PropListItem } from 'sys.ui.primitives/lib/ui/PropList';
 export { Card } from 'sys.ui.primitives/lib/ui/Card';
 export { BulletList } from 'sys.ui.primitives/lib/ui/BulletList';
 export { Text } from 'sys.ui.primitives/lib/ui/Text';
+export { ObjectView } from 'sys.ui.primitives/lib/ui/ObjectView';
 
 export {
   EventStack,
